% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesado.R
\name{add_from_data}
\alias{add_from_data}
\title{Registrar archivos de origen}
\usage{
add_from_data(
  raw_data_id,
  raw_data_path,
  from = data.frame(name = character(), id = character(), path = character())
)
}
\arguments{
\item{raw_data_id}{Espera un string con el id del drive del archivo de origen}

\item{raw_data_path}{Espera un string con el id del drive de la ubicacion del archivo de origen}

\item{from}{Espera un dataframe con los archivos de origen agregados antes, si es el primero lo crea vacio por default}
}
\value{
Retorna un dataframe que cada fila contiene un archivo de origen
}
\description{
Funcion que crea un dataframe con los archivos de los datos de utilizados para crear el dataframe
}
\details{
Primero obtiene el dribble del archivo en el drive.
Luego agrega una linea al dataframe con el nombre, id y ubicacion del archivo de origen.
Se debe utilizar de manera recursiva.
}
\examples{
from <- add_from_data(raw_data_id, raw_data_path)
from <- add_from_data(raw_data_id_2, raw_data_path_2, from)
}
