% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesado.R
\name{get_boundary_data}
\alias{get_boundary_data}
\title{Obtener los datos del perimetro del lote}
\usage{
get_boundary_data(file, raw_data_path, temp_path)
}
\arguments{
\item{file}{Espera un string con el nombre del archivo a descargar}

\item{raw_data_path}{Espera un string con el id de la ubicacion del archivo de origen}

\item{temp_path}{Espera un String con la ruta de la carpeta temporal donde se descargan y crean los distintos archivos}
}
\value{
Retorna un dataframe con el nombre del archivo, id de origen y la superficie
}
\description{
Funcion utilizada para obtener los datos del perimetro del lote cargado en el drive como gpkg
}
\details{
Primero obtiene los dribbles del archivo del perimetro.
Luego lo descarga y importa el archivo a un sf object.
Por ultimo arma un dataframe con el nombre del perimetro, id en el drive y la superficie.
}
\examples{
get_boundary_data(file, raw_data_path, temp_dir)
}
