% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacion.R
\name{get_idw_res}
\alias{get_idw_res}
\title{Obtener los Mejores Modelos de Predicción por IDW con Validación Cruzada}
\usage{
get_idw_res(
  path,
  file_name,
  sum_path,
  outlier = T,
  n_col = 1,
  seed = NULL,
  iqr_p = 3,
  nmax = 10,
  nmin = 2,
  nfold = 5,
  debug = 0
)
}
\arguments{
\item{path}{Ruta del directorio donde se encuentra el archivo GPKG.}

\item{file_name}{Nombre del archivo GPKG con los datos.}

\item{outlier}{valor lógico que define si sacar los outliers.
Por defecto es TRUE y no saca los outliers}

\item{n_col}{Número de columna a resumir en caso de tener múltiples columnas (e.g. datos apareados).
Por defecto, es 1.}

\item{seed}{Valor de la semilla a utilizar, si no se especifica utiliza la que queda por defecto.}

\item{iqr_p}{Si outlier es FALSE especifica que outliers filtrar.
Por defecto es 3}

\item{nmax}{Numero máximo de vecinos utilizados.
Por defecto es 10}

\item{nmin}{Numero mínimo de vecinos utilizados.
Por defecto es 2}

\item{nfold}{Numero de folds a utilizar en la validación cruzada.
Por defecto es 5}

\item{debug}{Este parámetro se utiliza en un entorno de prueba para cuando se quiera indicar que layers utilizar.
Puede ser un vector o un entero, por defecto es 0.}
}
\value{
Un dataframe con el mejor modelo de predicción por IDW y los datos de la validación cruzada para cada capa.
}
\description{
Esta función calcula y devuelve los mejores modelos de predicción por el método del inverso a la distancia (IDW) para cada capa utilizando validación cruzada.
}
\details{
La función setea el archivo del log.
Luego lista las capas del GPKG y arma la grilla de tuneo con diferentes valores de idp para cada capa.
Luego recorre la grilla de tuneo, para cada combinación de layer-idp obtiene la validación cruzada,
utilizando el método de k-folds y arma un dataframe donde cada fila tiene su combinación layer-idp y los resultados de la validación cruzada
Por ultimo calcula el error cuadrático medio de cada combinación layer-idp, filtra por capa las que tienen menor rmse y lo retorna
}
\examples{
get_idw_res(path="~/Documentos/Tesina/Data/clean", file_name="Elevacion.gpkg", seed = 701408733)
}
