% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacion.R
\name{get_ok_res}
\alias{get_ok_res}
\title{Obtener los Mejores Modelos de Predicción por OK con Validación Cruzada}
\usage{
get_ok_res(
  path,
  file_name,
  param,
  outlier = T,
  trainwo = F,
  iqr_p = 3,
  n_col = 1,
  seed = NULL,
  nfold = 5,
  debug = 0
)
}
\arguments{
\item{path}{Ruta del directorio donde se encuentra el archivo GPKG.}

\item{file_name}{Nombre del archivo GPKG con los datos.}

\item{param}{Dataframe con los parámetros iniciales para correr el modelo}

\item{outlier}{valor lógico que define si sacar los outliers.
Por defecto es TRUE y no saca los outliers}

\item{trainwo}{logico que determina si entrenar el modelo sin los outliers, Por defecto es FALSE}

\item{iqr_p}{Si outlier es FALSE especifica que outliers filtrar.
Por defecto es 3}

\item{n_col}{Número de columna a resumir en caso de tener múltiples columnas (e.g. datos apareados).
Por defecto, es 1.}

\item{seed}{Valor de la semilla a utilizar, si no se especifica utiliza la que queda por defecto.}

\item{nfold}{Numero de folds a utilizar en la validación cruzada
Por defecto es 5}

\item{debug}{Este parámetro se utiliza en un entorno de prueba para cuando se quiera indicar que layers utilizar.
Puede ser un vector o un entero, por defecto es 0.}
}
\value{
Un dataframe con el mejor modelo de predicción por OK y los datos de la validación cruzada para cada capa.
}
\description{
Esta función calcula y devuelve los mejores modelos de predicción por el método ordinary kriging (OK) para cada capa utilizando validación cruzada.
}
\details{
La función setea el archivo del log.
Luego carga en una grilla de tuneo los hiperparametros especificados en param.
Luego recorre la grilla de tuneo, obtiene la validación cruzada,
utilizando el método de k-folds y arma un dataframe donde cada fila tiene su layer, tiempo de ejecución y los resultados de la validacion cruzada y lo retorna
}
\examples{
get_idw_res(path="~/Documentos/Tesina/Data/clean", file_name="Elevacion.gpkg",param="~/Documentos/Tesina/Data/variograms/ECa30_vgm.RDS", seed = 701408733)
}
