% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacion.R
\name{get_prediction_model}
\alias{get_prediction_model}
\title{Obtener modelo de predicción}
\usage{
get_prediction_model(
  path,
  file_name,
  layer = NULL,
  method = NULL,
  idp = NULL,
  formula = NULL,
  dist_nn = NULL,
  knn = NULL,
  outlier = T,
  iqr_p = 3,
  n_col = 1
)
}
\arguments{
\item{path}{Ruta del directorio donde se encuentra el archivo GPKG.}

\item{file_name}{Nombre del archivo GPKG con los datos.}

\item{layer}{Nombre de la capa a obtener el modelo}

\item{method}{Metodo a aplicar para realizar la predicción.
Los valores posibles son "idw", "ok" y "rf"}

\item{idp}{Valor de potencia aplicar es idw}

\item{formula}{Formula de tendencia para aplicar ok}

\item{dist_nn}{Distancia promedio al vecino más cercano para aplicar idw y ok}

\item{knn}{Numero de vecinos a utilizar para aplicar rf}

\item{outlier}{valor lógico que define si sacar los outliers.
Por defecto es TRUE y no saca los outliers}

\item{iqr_p}{Si outlier es FALSE especifica que outliers filtrar.
Por defecto es 3}

\item{n_col}{Número de columna a resumir en caso de tener múltiples columnas (datos apareados).
Por defecto, es 1.}
}
\value{
Modelo de predicción según método especificado
}
\description{
Esta función calcula y devuelve el modelo de predicción del metodo y los parametros y layer especificado
}
