% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesado.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Obtener datos crudos}
\usage{
get_raw_data(raw_data_id, temp_path)
}
\arguments{
\item{raw_data_id}{Espera un string con el id del archivo de origen}

\item{temp_path}{Espera un String con la ruta de la carpeta temporal donde se descargan y crean los distintos archivos}
}
\value{
Retorna un dribble con la ruta local de los datos crudos
}
\description{
Funcion utilizada para obtener los datos crudos desde el drive
}
\details{
Obtiene el dribble del archivo de origen y lo descarga
}
\examples{
datos <- get_raw_data(raw_data_id,temp_dir)
}
