% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesado.R
\name{get_raw_shp}
\alias{get_raw_shp}
\title{Obtener archivos Shapefile}
\usage{
get_raw_shp(file, raw_data_path, temp_path)
}
\arguments{
\item{file}{Espera un string con el nombre del archivo a descargar}

\item{raw_data_path}{Espera un string con el id de la ubicacion del archivo de origen}

\item{temp_path}{Espera un String con la ruta de la carpeta temporal donde se descargan y crean los distintos archivos}
}
\value{
Retorna una lista de dos entradas con los datos y los archivos de origen
}
\description{
Funcion utilizada para obtener los datos crudos en formato shape desde el drive
}
\details{
Primero obtiene los dribbles del archivo de origen,
Luego lo descarga y arma el dataframe con la info de origen
Por ultimo importa el archivo a un sf object y arma la lista con el sf object y el df de origen
}
\examples{
datos <- get_raw_shp(file, raw_data_path, temp_dir)
}
