% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacion.R
\name{get_vgm_param}
\alias{get_vgm_param}
\title{Explorar variogramas}
\usage{
get_vgm_param(file_path, outlier = T, iqr_p = 3, n_col = 1, debug = 0)
}
\arguments{
\item{file_path}{Ruta del archivo GPKG con los datos.}

\item{outlier}{valor lógico que define si sacar los outliers.
Por defecto es TRUE y no saca los outliers}

\item{iqr_p}{Si outlier es FALSE especifica que outliers filtrar.
Por defecto es 3}

\item{n_col}{Número de columna a resumir en caso de tener múltiples columnas (e.g. datos apareados).
Por defecto, es 1.}

\item{debug}{Este parámetro se utiliza en un entorno de prueba para cuando se quiera indicar que layers utilizar.
Puede ser un vector o un entero, por defecto es 0.}
}
\value{
Un dataframe con el con la tendencia y los valores seteados para cada capa.
}
\description{
Esta función obtiene la tendencia que mejor R2 tenga grafica el variograma y permite setear los hiperparametros para utilizar en OK
}
\details{
Primero carga las capas de los datos indicados en la ruta.
Luego recorre capa por capa cargando los datos, obtiene la tendencia con mejor R2 y calcula el variograma.
Luego grafica el variograma, permite ingresar datos de sill, nugget y range, los grafica y pregunta si se quiere quedar con esos valores o cargar unos nuevos.
Por ultimo arma y retorna un dataframe con los valores seteados para cada capa
}
\examples{
get_vgm_param(file_path="~/Documentos/Tesina/Data/clean/Elevacion.gpkg",debug = 1)
}
