% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacion.R
\name{rfsi}
\alias{rfsi}
\title{Script for training RF spatial model}
\usage{
rfsi(obs, model = NULL, tgt = NULL, vname,  knn = 50, angle = T,  dist = T)
}
\arguments{
\item{obs}{SF with observed points}

\item{model}{Trained RFsi model}

\item{tgt}{SF with prediction grid or locations}

\item{vname}{Data variable name}

\item{knn}{Number of neighbors}

\item{angle}{Include angles}

\item{dist}{Include distances}
}
\value{
Object of class ranger or SF with prediction grid
}
\description{
RFsi is a function used to obtain models by the method Random Forest and to predict over a grid. Implemented using the ranger package (Wright & Ziegler, 2017).
}
\details{
If the model and grid are not specified, then the function obtains a model of class ranger and returns it.
On the other hand, if the model and the grid with the locations are specified, it returns a grid of class sf with the predictions.
}
\examples{
mod <- rfsi(obs, model = NULL, tgt = NULL, vname, knn = 50, angle = F, dist = T)
pred <- rfsi(obs, model = mod, tgt = grid, vname, knn = 50, angle = F, dist = T)
}
\author{
Carlos Agustín, Alesso
}
