% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpfca-package.R
\docType{package}
\name{interpfca-package}
\alias{interpfca}
\alias{interpfca-package}
\title{interpfca: Herramientas para el procesado, caracterización y modelado de interpolación de datos}
\description{
Paquete de funciones creadas para realizar el procesado, caracterización y modelado de interpolación de datos utilizados en la tesina.
}
\author{
\strong{Maintainer}: Alejandro Alemán \email{janoaleman@gmail.com}

}
\keyword{internal}
